# Speed up NewShaX