In many places we create a hash object, e.g. by calling `openssl.NewSHA256()`, just to get the type of hash and map it to an OpenSSL hash identifier. Then the original hash is discarded. These pattern is common enough to worth optimizing.

This PR refactors the `evpHash` implementation so that the hash `EVP_MD_CTX` context is created and initialized lazily. These are the speed gains:

```
goos: windows
goarch: amd64
pkg: github.com/golang-fips/openssl/v2
cpu: Intel(R) Core(TM) i7-10850H CPU @ 2.70GHz
             │    old.txt     │               new.txt               │
             │     sec/op     │   sec/op     vs base                │
NewSHA256-12   1533.50n ± 27%   50.39n ± 8%  -96.71% (p=0.000 n=10)

             │  old.txt   │            new.txt             │
             │    B/op    │    B/op     vs base            │
NewSHA256-12   24.00 ± 0%   24.00 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

             │  old.txt   │            new.txt             │
             │ allocs/op  │ allocs/op   vs base            │
NewSHA256-12   1.000 ± 0%   1.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```